# src/main/resources/swagger/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PlanetAPI Documentation",
    "description": "API for managing and querying planets in the Star Wars universe. It allows you to perform CRUD (create, read, delete) operations on planets, as well as access specific details such as name, weather, terrain and number of appearances in films. Integrates with the Star Wars public API to provide additional, up-to-date planet information.",
    "version": "1.22.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/planet": {
      "get": {
        "tags": [
          "planet-controller"
        ],
        "description": "Search all planets in database",
        "operationId": "getAll",
        "responses": {
          "200": {
            "description": "Return all planets in database",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanetDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "planet-controller"
        ],
        "description": "Add planet to database",
        "operationId": "addPlanet",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanetDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Unable to add planet"
          },
          "200": {
            "description": "Planet added successfully"
          }
        }
      }
    },
    "/planet/{id}": {
      "get": {
        "tags": [
          "planet-controller"
        ],
        "description": "Search for planet by id",
        "operationId": "findById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "No planet with this id was found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDTO"
                }
              }
            }
          },
          "200": {
            "description": "Return planet",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "planet-controller"
        ],
        "description": "Deletes planet to database",
        "operationId": "deletePlanet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Planet successfully deleted"
          },
          "404": {
            "description": "No planet with this id was found"
          }
        }
      }
    },
    "/planet/name/{name}": {
      "get": {
        "tags": [
          "planet-controller"
        ],
        "description": "Search for planet by name",
        "operationId": "findByName",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "No planet with this name was found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDTO"
                }
              }
            }
          },
          "200": {
            "description": "Return planet",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlanetDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PlanetDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "climate": {
            "type": "string"
          },
          "terrain": {
            "type": "string"
          },
          "appearedInFilms": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  }
}